Move addon manager files to addons/ folder
