+++ src/addons/addondemo.cpp
@@ -3,7 +3,7 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
 #include "addondemo.h"
-#include "addonmanager.h"
+#include "addons/manager/addonmanager.h"
 #include "leakdetector.h"
 #include "logger.h"
