+++ src/addons/manager/addondirectory.cpp
