+++ src/addons/manager/addondirectory.h
