+++ src/addons/manager/addonindex.cpp
