+++ src/addons/manager/addonindex.h
