+++ src/addons/manager/addonmanager.cpp
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-#include "addonmanager.h"
+#include "addons/manager/addonmanager.h"
 #include "leakdetector.h"
 #include "logger.h"
 #include "models/feature.h"
