+++ src/addons/manager/addonmanager.h
